requirements.txt
pystan==3.10.0
httpstan==4.13.0
numpy==1.26.4
pandas==1.5.3
appdirs==1.4.4
torch==2.1.1 --index-url https://download.pytorch.org/whl/cu118
triton==2.1.0
gluonts==0.15.1
torchmetrics==1.4.2
gpytorch==1.13
utilsforecast==0.2.5
causal-conv1d==1.4.0
mamba-ssm==2.2.2
matplotlib
seaborn